requirements.txt
-i https://pypi.org/simple
boto3==1.7.16
botocore==1.10.16
django-sass-processor==0.7
django-storages==1.6.6
django==2.0.5
docutils==0.14
jmespath==0.9.3
pillow==5.1.0
python-dateutil==2.7.2; python_version >= '2.7'
pytz==2018.4
s3transfer==0.1.13
six==1.11.0
